fix(paste): 修复目录迁移的 CI 识别
拆分窗口主题与同步模块并保留原文件主体，使 GitHub 将旧目录文件完整识别为迁移到 plugins/paste，避免构建任务继续解析已删除的 pasteboard-pro 目录。

AI-Co-Authored-By: Codex
