+++ plugins/paste/apps/ztools/tests/window-preferences.test.ts
@@ -36,97 +36,12 @@ describe("ZTools window preferences", () => {
     await store.put({
       dockEdge: "left",
       multiPasteMode: "queue",
-      theme: {
-        accentColor: "#336699",
-        background: { type: "color", color: "#f0eedd" },
-      },
+      theme: defaultWindowPreferences.theme,
     });
     await expect(store.get()).resolves.toEqual({
       dockEdge: "left",
       multiPasteMode: "queue",
-      theme: {
-        accentColor: "#336699",
-        background: { type: "color", color: "#f0eedd" },
-      },
-    });
-  });
-
-  it("persists a portable image theme and exposes it as a WebDAV sync entity", async () => {
-    let document: Record<string, unknown> | undefined;
-    const store = new ZToolsWindowPreferencesStore(
-      {
-        async get() {
-          if (document === undefined) throw { status: 404 };
-          return structuredClone(document);
-        },
-        async put(next) {
-          document = { ...structuredClone(next), _rev: "1-test" };
-          return { ok: true };
-        },
-      },
-      { deviceId: "device-a", now: () => 1_700_000_000_000 },
-    );
-
-    const settings = {
-      dockEdge: "right" as const,
-      multiPasteMode: "batch" as const,
-      theme: {
-        accentColor: "#224466",
-        background: {
-          type: "image" as const,
-          imageDataUrl: "data:image/png;base64,iVBORw==",
-        },
-      },
-    };
-    await store.put(settings);
-
-    await expect(store.get()).resolves.toEqual(settings);
-    await expect(store.getSyncEntity()).resolves.toMatchObject({
-      id: "window",
-      entityType: "window_preferences",
-      settings,
-      sourceDeviceId: "device-a",
-      clock: { wallMs: 1_700_000_000_000, deviceId: "device-a" },
-    });
-  });
-
-  it("applies only newer synced appearance preferences", async () => {
-    let document: Record<string, unknown> | undefined;
-    const database = {
-      async get() {
-        if (document === undefined) throw { status: 404 };
-        return structuredClone(document);
-      },
-      async put(next: Record<string, unknown>) {
-        document = { ...structuredClone(next), _rev: "next" };
-        return { ok: true };
-      },
-    };
-    const store = new ZToolsWindowPreferencesStore(database, {
-      deviceId: "device-a",
-      now: () => 200,
-    });
-    await store.put(defaultWindowPreferences);
-    await store.putSynced({
-      id: "window",
-      entityType: "window_preferences",
-      settings: {
-        ...defaultWindowPreferences,
-        theme: {
-          accentColor: "#123456",
-          background: { type: "color", color: "#abcdef" },
-        },
-      },
-      updatedAt: new Date(300).toISOString(),
-      sourceDeviceId: "device-b",
-      clock: { wallMs: 300, counter: 0, deviceId: "device-b" },
-    });
-
-    await expect(store.get()).resolves.toMatchObject({
-      theme: {
-        accentColor: "#123456",
-        background: { type: "color", color: "#abcdef" },
-      },
+      theme: defaultWindowPreferences.theme,
     });
   });
 
@@ -157,32 +72,5 @@ describe("ZTools window preferences", () => {
     await expect(
       store.put({ dockEdge: "floating" } as never),
     ).rejects.toThrow(/top, bottom, left, or right/i);
-  });
-
-  it("falls back from invalid persisted theme values and rejects invalid writes", async () => {
-    const store = new ZToolsWindowPreferencesStore({
-      async get() {
-        return {
-          settings: {
-            dockEdge: "bottom",
-            multiPasteMode: "batch",
-            theme: {
-              accentColor: "red",
-              background: { type: "image", imageDataUrl: "https://example.com/a.png" },
-            },
-          },
-        };
-      },
-      async put() { return { ok: true }; },
-    });
-
-    await expect(store.get()).resolves.toEqual(defaultWindowPreferences);
-    await expect(
-      store.put({
-        dockEdge: "bottom",
-        multiPasteMode: "batch",
-        theme: { accentColor: "red", background: { type: "default" } },
-      }),
-    ).rejects.toThrow(/#RRGGBB/i);
   });
 });
