+++ plugins/paste/pnpm-workspace.yaml
@@ -4,8 +4,3 @@ packages:
 
 allowBuilds:
   esbuild: true
-
-overrides:
-  brace-expansion: 2.1.4
-  nanoid: 3.3.18
-  postcss: 8.5.23
